Let me start this thing. I am just slightly over my grogginess.
Yeah, I regret it. The ideal conclusion to this would be to make Spiral pluggable rather than redo it in Pharo.

It is obvious to me now that making Spiral truly useful will require sustained effort over the next few years.

It will be long war.

I will not make a mistake in directing my efforts again. It would have made more sense to go with a dynamic language back in 2015. But that it just hindsight. Had I this knowledge back in 2015, Spiral would have long been done by now.

I made a mistake thinking that Spiral could be a real language rather than a DSL. I got tricked by its many powerful features.

I know that C++ is a real language and saw how Spiral exceeds it in every way so I thought that Spiral is an even realer language. This was a mistake.

Spiral's true form is a performance DSL.

The truth is, I really do want to redo it in Pharo, but I am worried about the lack of pattern matching in that language.

It is troublesome. Pharo could stand to be a more complex language. But I think that even as it is I could make it work somehow.

Well, it is important to grab a foothold.

Deep learning is just too loose. CFR will just have to do.

I regret it. I just haven't gotten the reward that I sought for all the long hours I've put into programming. How great it would be if I could do one month of work in a single day? Where would I be now if that was the case? It sucks being a human.

But this is what I desired and I knew it would be like this when I set out on this path.

Eventually Spiral will become my power. At some point I will be done with it and that will be a cause to celebrate. That will have made it a real achievement.

I really will do in Pharo pattern matching or not when I am done with all of this, but I'll save that work for much later, maybe 2020.

I need to get more familiar with that language before I try undertaking something massive like porting Spiral to it.

Well, it is not that massive - 4k LOC is still a month of work at best even in Pharo, but I want to take a break from Spiral.

And before I can take a break I need to finish Spiral first.

Today the theme will be dealing with error message. Let me finally finish the parser so I can start the testing. I can't believe I've started thinking of it as a fun thing. That is how tired I am of the parser.