Let me just finish chilling and then I will start.
I regret it. I wish there was a better way than going the programming route. Life is something that should be beatable using standard gaming tactics. Maybe for other people it is, but not for me.

Losing a game is one thing, but losing to those you look down on is bound to shatter your self conception.

Programming is the grimm trigger. Programming is not the kind of thing you do to your friends.

But this world is not my friend.

And I intend to drive my enemies to absolute extinction. My policy will not be to take lives, but their livelihoods and drive them to the deepest, darkest pits of despair. They brought it on themselves by making me lose too much.

We'll see how cocky they are once I start making cracks in the credit assignment problem.

Eventually programming will stop being the drudge work that I am doing now.