"1:50pm.
```
Let
  (18,Op (3,ListCreate,[|Lit (1,LitInt64 2L); Lit (2,LitInt64 3L)|]),
   ListTakeAllTest
     (17,2,0,
      ListTakeAllTest
        (16,0,1,
         ListTakeAllTest
           (15,0,2,MoveLocalPtrTo (5,3,Op (4,ListCreate,[||])),
            MoveLocalPtrTo
              (14,3,
               Let
                 (13,
                  Function (7,Function (6,Op (4,ListCreate,[||]),[||],1),[||],1),
                  Op
                    (12,Apply,
                     [|Op (10,Apply,[|V (8,3); Lit (9,LitInt64 4L)|]);
                       Lit (11,LitInt64 5L)|])))),
         MoveLocalPtrTo
           (14,3,
            Let
              (13,Function (7,Function (6,Op (4,ListCreate,[||]),[||],1),[||],1),
               Op
                 (12,Apply,
                  [|Op (10,Apply,[|V (8,3); Lit (9,LitInt64 4L)|]);
                    Lit (11,LitInt64 5L)|])))),
      MoveLocalPtrTo
        (14,3,
         Let
           (13,Function (7,Function (6,Op (4,ListCreate,[||]),[||],1),[||],1),
            Op
              (12,Apply,
               [|Op (10,Apply,[|V (8,3); Lit (9,LitInt64 4L)|]);
                 Lit (11,LitInt64 5L)|])))))
```

Yeah, this works now. The print out above makes sense too. I think I got it right on the first try. Go me.

With this, let me continue the testing. I should be nearly done with the regular tests at this point.

```
Unhandled Exception: System.IndexOutOfRangeException: Index was outside the bounds of the array.
   at Spiral.PartEval.push_var(TypedData x, LangEnv d) in C:\Users\Marko\Source\Repos\The Spiral Language\The Spiral Language\SpiralPartEval.fs:line 197
   at Spiral.PartEval.partial_eval(LangEnv d, Expr x) in C:\Users\Marko\Source\Repos\The Spiral Language\The Spiral Language\SpiralPartEval.fs:line 588
```

Of course as soon as I put in the Core library I get this.

```
vartag=63

Unhandled Exception: System.IndexOutOfRangeException: Index was outside the bounds of the array.
   at Spiral.PartEval.push_var(TypedData x, LangEnv d) in C:\Users\Marko\Source\Repos\The Spiral Language\The Spiral Language\SpiralPartEval.fs:line 197
```

Hmmm, this is not quite what I expected.

2:10pm.

```
x=4
x=6
x=12
x=12
x=16
x=17
x=17
vartag=63
```

These are all the subrenamed vartags.

Then how come the one case in the test has been left alone. I am super confused here.

Ok, I can confirm that the one in the test is not getting subrenamed.

Why?

2:15pm.

```
vartag=63
len=4

Unhandled Exception: System.IndexOutOfRangeException: Index was outside the bounds of the array.
   at Spiral.PartEval.push_var(TypedData x, LangEnv d) in C:\Users\Marko\Source\Repos\The Spiral Language\The Spiral Language\SpiralPartEval.fs:line 197
   at Spiral.PartEval.partial_eval(LangEnv d, Expr x) in C:\Users\Marko\Source\Repos\The Spiral Language\The Spiral Language\SpiralPartEval.fs:line 591
```

This is the size of the local stack.

The `Core` functions, are they in the local or the global stack?

```
        |> timeit env.timing.prepass (Prepass.prepass {prepass_context=context; prepass_map=env.map; prepass_map_length=count})
    let module_ =
        let d = {rbeh=AnnotationDive; seq=env.seq; env_global=context; env_stack_ptr=0; env_stack=Array.zeroCreate size; trace=[]; cse=ref Map.empty}
        //printfn "%A" expr
        timeit env.timing.peval (PartEval.partial_eval d) expr
```

Well it definitely can't be in the local stack.

2:20pm. Damn the IDE is being crappy. It has started to really grind to a halt. I get the feeling that there is a memory leak in it somewhere.

2:25pm. Ok, I know what to do. I've been considering calling the subrenaming in the `module_let`, but that would be too complicated. What I will do instead is move the global ptr.

```
| MoveGlobalPtrTo(tag,vartag,on_succ) -> MoveGlobalPtrTo(tag,rename vartag,f on_succ)
```

This is how it should be. Now...

2:30pm.

```
| MoveGlobalPtrTo(_,vartag,on_succ) -> ev {d with env_stack_ptr=vartag - d.env_global.Length} on_succ
```

This things can turn out to be subtle and complicated.

Let me give it a shot. I think it should work now.

```
{a=lit 4i64; b=lit 5i64}
```

Yeah, now it works as expected. So much for getting it right on the first try. Actually the idea of doing it like this occurred to me, but I went with the wrong one for no reason.

This issue really took me a while.

Let me finally complete the test.

2:30pm.

```
Unhandled Exception: System.NullReferenceException: Object reference not set to an instance of an object.
   at Spiral.Types.TypedData.get_Tag()
   at Spiral.PartEval.partial_eval(LangEnv d, Expr x) in C:\Users\Marko\Source\Repos\The Spiral Language\The Spiral Language\SpiralPartEval.fs:line 610
   at Spiral.PartEval.partial_eval@610-10.Invoke(LangEnv d, Tuple`2 _arg1) in C:\Users\Marko\Source\Repos\The Spiral Language\The Spiral Language\SpiralPartEval.fs:line 610
```

I guess I was wrong again.

2:35pm. Agh, this damn test. I need to debug it again. Let me do so.

Ok, let me do it from `test8`.

```
inl option_int = .Some, (1,2,3) \/ .None

inl x = join Type (box: .None to: option_int)
match x with
| #(.Some, x) -> x
| #(.None) -> 0,0,0
```

```
inl #x = join Type (box: .None to: option_int)
match x with
| .Some, x -> x
| .None -> 0,0,0
```

Like this it works.

```
    | MoveGlobalPtrTo(_,vartag,on_succ) ->
        //printfn "vartag=%i" vartag
        //printfn "d.env_global.Length=%i" d.env_global.Length
        //ev {d with env_stack_ptr=vartag - d.env_global.Length} on_succ
        ev d on_succ
```

How could this be happening. I get the same error even if I do this.

```
    let move_ptr = id //if clauses.Length > 1 then RawMoveGlobalPtrTo else id
    Array.foldBack (fun (pat, exp) on_fail -> cp arg pat (move_ptr exp) on_fail) clauses (op ErrorPatMiss [|v arg|])
```

Or this. I did something deeply wrong. Did I introduce an error while swapping the position of `on_succ` and `on_fail` in the prepass?

2:45pm. This error is definitely more long winded than I thought it would be.

Let me backtrack to an earlier commit and I will do the changes from the top one by one."