"``` Offset and length were out of bounds for the array or count is greater than the number of elements from index to the end of the source collection. ```
Sometimes I get this error when I try to switch branches. Now it is hitting me just where it hurts. I'll try updating the IDE.

2:50pm. I guess I'll take a short break while it updates.

https://mangadex.org/title/20573/machigatta-ko-wo-mahou-shoujo-ni-shite-shimatta

This came out too.

3pm. Agh, let me do the chores. The download is 2Gb.

3:40pm.

```
Offset and length were out of bounds for the array or count is greater than the number of elements from index to the end of the source collection.
```

FFFFFFFFuck the IDE!

Let me see what the problem is.

3:45pm. How can I possibly see!?

Today is a bust. Basically what I have to do is switch to a better IDE. Maybe VS Code?

No way am I going to resolve this that easily.

Uah...

3:50pm. I am just groaning here right now.

Let me ask on SO.

https://stackoverflow.com/questions/54789293/how-do-i-deal-with-the-array-out-of-bounds-exception-when-creating-a-new-git-bra

4pm. I'll try switching from the command line and if that fails, switching the IDE."