"``` type SpiralType0 =     | SpiralType0_1 of int64 * int64     | SpiralType0_2 of int64 let rec method_0 () : SpiralType0 =     SpiralType0_2 1L let ((var_1 : SpiralType0)) = method_0 () let ((var_2 : SpiralType0)) = method_0 () let ((var_18 : bool)) =     match var_1 with     | SpiralType0_1 (var_3, var_4) ->         match var_2 with         | SpiralType0_1 (var_5, var_6) ->             let ((var_7 : bool)) = var_3 = var_5             if var_7 then                 var_4 = var_6             else                 false         | SpiralType0_2 var_10 ->             false     | SpiralType0_2 var_12 ->         match var_2 with         | SpiralType0_1 (var_13, var_14) ->             false         | SpiralType0_2 var_15 ->             var_12 = var_15 ```
And work it does. Phew. As it turns out, I've definitely been overeager with memoization in places. Before this it was the `term_vars` function.

6:05pm. I am going to call it a day here.

There was one thing I forgot namely the or patterns. But solitary or patterns are pretty rare and even more so those with large bodies. Have I ever actually seen a solitary or pattern with a large body? I do not think so.

I do not think I have to bend myself over for it here.

I am going to call it a day here rather than continuing on with the tests.

This bug took me pretty much the entire to deal with. It really kicked my ass. I wanted to step on a bug, but it turned out to be a land mine. That is some roach!

Honestly, I do not think I could have avoided it. It would have taken some truly exceptional reasoning to anticipate something like this ahead of time. There were like 3 very subtle issues packed into one bug.

6:15pm. I am actually quite pleased with how the day went. I cleared out most of the remaining regular tests, really beefed up the partial evaluation in the `If` op and now resolved these binding issues. Thanks to this last one, I can now be much more certain that no big issues are lurking in the binding mechanisms. I was expecting something here at some point. I knew that indexing into variables using arrays would have surprises lying around as that sort of thing is complex enough that I haven't been able to fully reason it out.

I just trusted myself to deal with whatever would come up.

I am done for the day."