"6:45pm. Done with lunch. Right now I've playing with union types for a bit to see if I can trigger an error somehow, but no luck.
```
inl f x =
    match x with
    | q,w,e,r,t,z,x,c,v,b,m -> 0
    | (({q w e r t y z a b}, _, _) | ((),a,b)) when a > 0 ->
        inl f a b = a + b
        f a b
    | a,b -> a + b
dyn (f ((),2,3))
```

No, nope. It is probably solid at this point.

...

```
inl f x =
    match x with
    | q,w,e,r,t,z,x,c,v,b,m -> 0
    | (((),a,b) | ({q w e r t y z a b}, _, _)) ->
        inl f a b = a + b
        f a b
    | a,b -> a + b
dyn (f ({q=(); w=(); e=(); r=(); t=(); y=(); z=(); a=1; b=2},2,3))
```

Ah, right. I see it.

```
        | RawMoveGlobalPtrTo x ->
            let on_succ,free_vars,stack_size = memoize prepass_memo_dict (loop env) x
            if stack_size = 0 then on_succ,free_vars,stack_size
            else MoveGlobalPtrTo(tag(),env.prepass_map_length,on_succ),free_vars,stack_size
```

I kind of did not think it through completely, but I am not memoizing the `RawMoveGlobalPtrTo` themselves, but their bodies. That is why this thing works. It is sound completely.

There is not a single thing I need to fear. The construct will hold under stress.

Also I've just realized that the prepass itself will tell me if I try an union pattern with one of the branches unbound.

That is cool.

6:55pm. https://mangadex.org/title/21119/satanophany

This came out too.

```
    let move_ptr = id //if clauses.Length > 1 then RawMoveGlobalPtrTo else id
    Array.foldBack (fun (pat, exp) on_fail -> cp arg pat (move_ptr exp) on_fail) clauses (op ErrorPatMiss [|v arg|])
```

Whops, forgot to put this in. It turns out I haven't been testing it with memoization at all. No wonder it was solid.

```
    let inline f move_ptr = Array.foldBack (fun (pat, exp) on_fail -> cp arg pat (move_ptr exp) on_fail) clauses (op ErrorPatMiss [|v arg|])
    if clauses.Length > 1 then f RawMoveGlobalPtrTo else f id
```

Let me do it like so.

```
Unhandled Exception: System.IndexOutOfRangeException: Index was outside the bounds of the array.
   at Spiral.PartEval.partial_eval@584-2.Invoke(Int32 x) in C:\Users\Marko\Source\Repos\The Spiral Language\The Spiral Language\SpiralPartEval.fs:line 584
   at Spiral.PartEval.partial_eval(LangEnv d, Expr x) in C:\Users\Marko\Source\Repos\The Spiral Language\The Spiral Language\SpiralPartEval.fs:line 584
   at Spiral.PartEval.partial_eval(LangEnv d, Expr x) in C:\Users\Marko\Source\Repos\The Spiral Language\The Spiral Language\SpiralPartEval.fs:line 588
```

Nooooooooooooo!!!

7:05pm. Ah, I'll take care of it tomorrow. This error is happening in `Core` for some reason."