"8:20am. I am up. I still can't even begin to imagine how a function that is not even applied could result in a bug during partial evaluation, but it does not matter.
I've had a lot of time to think things through more dispassionately during the night and I've concluded that my memoization efforts during the prepass were sheer lunacy. I thought my grip on memoization was getting stronger, but it obvious was not as strong as I thought it was.

I was completely wrong that I could only memoize the expression and not the the environment associated with it. That was just so bad of me. But even if for the sake of the argument I assumed that was true, there is one feature in Spiral which firmly blocks that.

```
inl (a,() | a,Core) =
    open Core
    ...
```

The modules pretty much impose the requirement that I have to walk the node each time. Forget memoization of expressions in its presence.

8:35am. Let me chill a bit and then I will start.

What I will start is with redoing the prepass. I have a bunch of good ideas that should hopefully resolve all of this crap. I am going to make...no wait, I can't make modules dynamic what am I thinking?

8:40am. Ahhh, I'll do something at any rate. The correct thing to do is in the neighborhood of what I am thinking now. Let me chill a bit and then I will start.

Yeah, figuring out the correct semantics is hard, but it will be doable. I'll get the prepass down and there will be no more of these crazy binding errors."