"9:35am. I've finished chilling a while ago and have just been thinking.
I am trying to figure out a way to deal with `open`. I never expected this single thing to be such a huge problem.

9:40am. The ability to have arbitrary `open`s in the language is the source of all of its distortions. The prepass memoization would be a lot easier without it. Right now it is already a shell of its former self.

What I will do is move it outside the language, to the module declarations. I'll put the `open`s straight into `SpiralModule`. That will take care of it.

9:50am. https://www.reddit.com/r/ProgrammingLanguages/comments/at0alm/mlir_primer_a_compiler_infrastructure_for_the_end/

I'll go for this later. Forget about that for now.

It think it is time that I rev up my engine and get going. I can do this. After this change there will be nothing stopping me from finishing the language. It will be cast in iron and clad in stone. To be honest, apart from doing what I am going to do now since without memoization the prepass works perfectly, just leaving it like this is also an option I've considered. Maybe I would have even taken it if I was not in the middle of a redesign.

I can do better than this. There is no need to stop with just what I have now. I can keep going.

10:15am. I've really gotten into those slides by Latner.

I'll start once I am done with them.

10:25am. Let me start. This will take a me a few days. The prepass is nothing special, in fact it will be even simpler than before, but this will lead to a cascade of changes and I will have to go through the tests from the top. Let me start by removing that `MovePtr` garbage. I'll also remove the `Open` cases.

The slides were interesting. It feels like we are on the same page on what work is necessary, though they are coming at it from a different angle. I guess getting hired to work on that stuff is an option if my computer ever gives out.

But let me do what I can with my own power for now."