"10:40am.
```
    member d.ReadVar =
        let pos = d.Pos
        d.TryCurrent <| function
            | TokVar(_,t') ->
                let _ = memoize' var_position_dict (fun x -> x,pos) t'
                d.Skip; Ok t'
            | _ -> d.FailWith(ExpectedVar)

    member d.ReadOp =
        let pos = d.Pos
        d.TryCurrent <| function
            | TokOperator(_,t') ->
                let _ = memoize' var_position_dict (fun x -> x,pos) t'.name
                d.Skip; Ok t'
            | _ -> d.FailWith(ExpectedOperator')

    member d.ReadOpAsVar =
        let pos = d.Pos
        d.TryCurrent <| function
            | TokOperator(_,t') ->
                let _ = memoize' var_position_dict (fun x -> x,pos) t'.name
                d.Skip; Ok t'.name
            | _ -> d.FailWith(ExpectedOperator')
```

There is a certain amount of nastiness to this, but I can't disagree that this is an effective way to do extensibility.

10:45am.

```
// Note: The VarStrings are annotated with positional information using `var_position_dict` global in ParserCombinators.
```

Let me add this note. Let me just say that I have checked and the WCT is in fact thread safe.

Now let me modify the prepass so sets are used instead of that Map.

The main mistake that I did with the idea is to not jack the parser combinators.

Actaully, I should do it even earlier than that and do it in the tokenizer.

11am.

```
            try
                let op = op name
                let _ = memoize' var_position_dict (fun _ -> s.UserState,start.line,start.column) name
                Reply(TokOperator(pos,op))
            with :? TokenizationError as x -> Reply(Error, messageError x.Data0)
```

Here is the thing in Op.

Now let me convert the Prepass.

11:15am.

```
    let free_vars, stack_size = free_vars_and_stack_size x
    let free_vars =
        let er x = failwithf "Compiler error: The variable `%s` must be annotated with its position in the tokenizer." x
        Set.toArray free_vars
        |> Array.map (fun x -> x, memoize' Tokenize.var_position_dict er x)
    renaming x, free_vars, stack_size
```

Wonderful. Now the `free_vars` have their positional information where they should.

Let me clean up the parser. I am going to add compiler settings to the `ParserEnv` to make things more consistent.

11:20am. Ok, done with the parser cleanup. Now everything is consistent.

Let me get all the type errors out of the way in the compiler. After that is done, I will be able to resume testing.

```
    context : ResizeArray<TypedData>
    map : Map<string, int>
```

Let me merge these two.

11:30am.

```
let module_open (env: ModulePrepassEnv) x =
    match env.map.TryFind x with
    | Some x ->
        match x with
        | TyMap x -> {env with map=Map.foldBack (keyword_to_string >> Map.add) x env.map}
        | x -> raise_compile_error <| sprintf "Expected as module in `module_open`.\nGot: %s" (show_typed_data x)
    | _ -> raise_compile_error <| sprintf "In module_open, `open` did not find a module named %s in the environment." x
```

I think this should be all of them. I just need to do the printing as well.

```
        let unbound_variables =
            Array.choose (fun (name,pos) ->
                match Map.tryFind name env.map with
                | Some x -> None
                | None -> Some(name,pos)
                ) free_vars
```

11:40am.

```
    let module_ =
        let unbound_variables =
            Array.choose (fun (name,pos) ->
                match Map.tryFind name env.map with
                | Some x -> None
                | None -> Some(name,pos)
                ) free_vars
        match unbound_variables with
        | [||] -> ()
        | [|var,pos|] -> raise_compile_error' [pos] <| sprintf "The variable %s is unbound." var
        | _ ->
            let var, pos = Array.unzip unbound_variables
            let var = String.concat ", " var
            raise_compile_error' (Array.toList pos) <| sprintf "The variables %s are unbound." var
        let context =
            Array.map (fun (name,pos) ->
                match Map.tryFind name env.map with
                | Some x -> x
                | None -> failwith "impossible"
                ) free_vars
```

This is how it should be done.

Despite me taking plenty of time to think about how I want to propagate positional information last night, somehow it took me a really long time to get it done despite how simple the thing I did in the end turned out to be.

11:45am. I am still a bit on the fence about this kind of use of global dictionaries, but it should be fine. The code is certainly simpler and more elegant this way than how it would be otherwise. So it is fine.

And the prepass is great right now, though it is complex enough piece of work that I am going to go through the tests from the start. But assuming my general plan is right, then the issue of pattern compilation is as good as closed. That reminds me.

11:55am.

```
            let inline g f = Array.foldBack (fun (pat, exp) on_fail -> cp arg pat (f exp) on_fail) clauses (op ErrorPatMiss [|v arg|])
            if clauses.Length > 1 then g (fun x -> ap (func "" x) B) else g id
```

This should do the trick.

I am finally done with this stuff.

From here on out, it is just testing. Let me take a break here."