Done with the macro section. Let me plug the changelog.
I will stop here for a bit. The next will come arrays and the rest of the libraries. Also, I need to figure out what is left in the old `Extern` module. I think `show` was there. That needs its own module.

Well, none of this is going to get done quickly. The work I have now is straightforward, but there is a lot to through. I just intend to do enough to get to the point where I can be confident in the integrity of the language.