"5:35pm. Let me start work on the parsing library. I'll go over it just once.
```
// Primitives
inl m x = {
    elem =
        match x with
        | {parser_rec} {d with on_type} state -> join (parser_rec d .elem d state : on_type)
        | {parser} -> parser
        | {parser_mon} -> parser_mon .elem
    }
inl term_cast p typ = m {
    parser = inl d state ->
        p .elem {d with
            on_succ =
                inl k = term_cast (inl x,state -> self x state) (typ,state)
                inl x state -> k (x,state)
            } state
    }
```

I am not going to even try refactoring this.

```
// TODO: Instead of just passing the old state on failure to the next parser, the parser should
// compare states and fail if the state changed. Right now that cannot be done because Spiral is missing
// polymorphic structural equality on all but primitive types. I want to be able to structurally compare anything.
```

For how long is this going to be here?

```
inl convert_type = Macro fs [text: "System.Convert"]
inl to_int64 x = FS.StaticMethod convert_type .ToInt64 x int64
```

I have unbuilts for this sort of thing now.

```
inl sprintf format =
    inl strb_type = fs [text: "System.Text.StringBuilder"]
    inl strb = FS.Constructor strb_type (64i32)
    inl append x = FS.Method strb .Append x strb_type |> ignore
    sprintf_template append {
        on_succ = inl x _ -> x
        on_fail = inl msg _ -> FS.Method strb .ToString() string
        } format
```

This I need to refactor. I'll put `StringBuilder` into its own module.

5:45pm. Ok, I went over it and did some touching up. I am not sure I feel like messing with the parsing tests now.

Tomorrow I will deal with this and the `hacker_rank` tests.

https://arxiv.org/abs/1902.10250
Diagnosing Bottlenecks in Deep Q-learning Algorithms

I'll read this thing later.

5:55pm.

```
We believe the best strategy is to keep large architectures but carefully select the
number of gradient steps used per sample. We showed that
employing oracle early stopping techniques can provide
huge benefits in the performance in Q-learning. This motivates the future research direction of devising early stopping
techniques to dynamically control the number of gradient
steps in Q-learning, rather than setting it as a hyperparameter
as this can give rise to big difference in performance.
```

I am just skimming this thing, but this is what I've been thinking last night. Dynamically controlling the gradient steps would actually move deep learning closer to being Bayesian. Right now one of the huge issues would be that deep learning does not distinguish between one sample given 100 times and 100 identical samples given one time. There is a large difference from a Bayesian perspective between the two.

```
Surprisingly, we do not find on-policy distributions to be the most performant, but rather methods which
have high state-entropy and spread mass uniformly among
state-action pairs, seem to be highly effective for training.
Based on these insights, we propose a new weighting distribution which balances high-entropy and state aliasing, AFM,
that yields fair improvements in both tabular and continuous
domains with state-of-the-art off-policy RL algorithms.
```

Yeah, this is what I should be doing. This kind of science. I would do this sort of thing for any normal program, so why haven't I done it for deep RL?

6pm. Ah, I can't be assed to read this properly. I am definitely going to deal with the parsing stuff tomorrow.

A good case could be made for simply dropping the parsing stuff from the testing suite as the tests are overly large in terms of code generated, but I do not dare greenlight the language until I've gone through this.

I might consider dropping it in favor of something else later on after I am done with this, but not now.

6:05pm. I will continue with the plan and do all the tests as intended. The parsing library should be refactored at some point, but not now. It just has such a massive amount of partial application in it and would be a nightmare to refactor without type inference like in F#. With the new keyword arguments I would have designed it much differently now. It is the ugly as sin, but it is also a great stressor for the language. This is exactly why I need.

Keep it up me. Just 2-3 more days of this and I will be free. I am definitely looking forward to that.

I am on fire. The 4k LOC of Spiral is nothing. Assuming I have a good template for it, I can port it to other languages in no time at all.

This means I have absolute power to command the hardware and need not fear the bottlenecks of dynamic languages anymore. The programming world is my oyster.

I need to dig deep and master the aspects of virtual machines. Pharo is an ideal target for that. At some point I might even look into aspects of extensional typing. After all the practice I got on Spiral this feels like a piece of cake, but at the same time that is why this is not that important. It is just a compiler pass.

6:20pm. I really am looking forward to the future. I will jack my brain into the virtual machine by the end of all of this.

Spiral right now is better than ever.

And a few years of more effort like this, so will my skills be at unfathomable heights."