"7:35am. I am going to stop testing here. I am going to do the Queue so I can test the heap mutable types, but apart from that I won't bother with anything else. Actually, let me just make a test for heap mutables.
```
Error trace on line: 88, column: 21 in module Core.
    set:field:to: = !SetMutableRecord(set,field,to)
                    ^
The record's field and the value being set do not have the same type.
Got: var (int64)
And: lit 3i64
```

Oh the error message is excellent this time.

7:45pm.

```
inl m = heapm {a=type ""; b=1; c=heap {a=dyn 1; b=type ""}}
Record set: m field: .c to: heap {a=dyn 2; b=type ""}
```

This just works. I really like Spiral's new syntax tools. Keyword arguments are a bliss.

Now, let me finish this round of testing.

```
The timings are: {parse = 00:00:01.3259925;
 prepass = 00:00:00.3602247;
 peval = 00:00:00.2478365;
 codegen = 00:00:00.1198841;}
The time it took to run all the tests is: 00:00:02.0539378
```

My sense is that the compiler is quite a bit faster than it was last time. Even if I did the rest of the tests, last time was dominated by parsing regardless of that. And the rest even when put together are still only a third of parsing. So I'd get 2.5s for the full suite vs 9s last time.

This is pretty great. I think that 4x speedup would have been worth it. If anything on the ML library tasks, I'd expect a speedup in excess of 10.

7:50am. What comes now is conviction."