I am definitely finding it hard to cope with Pharo's simplicity, especially doing things without a lack of pattern matching or inbuilt datatypes like tuples. I've decided to put `regret` directly into instances of actions, but now I am not sure how to proceed here.
Where do I put the `strategySum`? When do I update the actions? Because of the lack of my usual tools, I find that my thinking is scrambled. This is going to be tougher than it looks at first glance. I could do this in 10m in either Spiral or F#, but here it might take me as much as few days. I should ask around on the Pharo discord or something.

Let me open a script in the repo so I can jot down my thoughts there. I am already getting withdrawal from this. I'll do the RPS example in F# before asking around how to do it in Pharo. I'll probably ask how to do it in Pharo even after doing it independently just to see if I've missed something design-wise.