"5:05pm.
```
 size := Action subclasses size.
 regretSum := (Array new: size) collect: [ :each | 0.0 ].
 strategySum := (Array new: size) collect: [ :each | 0.0 ].
```

How do I initialize an array with constants? This can't be the way to do it.

https://gist.github.com/jdevoo/8e8866cd6087e05790841d0f20b2e377

5:20pm. Agh, I have no idea. Let me just loop it.

I'll ask around later as to how to do this.

```
 regretSum := Array new: size. "How do I zero init the arrays?"
 strategySum := Array new: size.
 size timesRepeat: [ :i |
  regretSum at: i put: 0.0.
  strategySum at: i put: 0.0 ]
```

Lame.

5:40pm. Actually, I've caught fire and things are going well. I put those arrays into a `Regret` class. This makes initing them easier.

Let me start making a list of things to ask about Pharo:
- Tab indents, but Shift+Tab does not dedent unlike in all the other popular editors. Why?
- Where is map2, mapFold and such?

6:20pm. Done with lunch. I am tempted to just call it a day here, but let me give making action a shot. I'll actually test it tomorrow.

6:25pm.

```
action: strategy
 strategy inject: 0.0 into: [ :prob :each |

  if prob < each
   ]
```

This is really so annoying. The lack of pattern matching is really killing the expressiveness of the language. I prefer the functional approach after all, but it is not like this can't be made to work.

6:45pm.

```
action: strategy
 |prob r actions |
 r := SharedRandom globalGenerator next.
 prob := 0.0.
 actions := Action subclasses.
 actions size timesRepeat: [ :i |
  prob := prob + strategy at: i.
  prob < r ifTrue: [^actions at: i].
  ].
 self error.
```

This has some safety issues, but I'll roll with it. I think I understand why map2 is not there - because all the maps are directly used on the object.

Do I have anything else I am missing?

7:05pm.

```
myAction := strategy action.
```

Agh, this thing is really not working for me. I think I am almost done with the the program, but it is not going particularly well. Let me put `action` into `Action class >> sample:`

7:10pm.

```
normalized
 | strategy normalizingSum |
 strategy := sum collect: [ :each | each max: 0.0 ].
 normalizingSum := strategy sum.
 ^strategy collect: [ :each |
  normalizingSum strictlyPositive
  ifTrue: [ each / normalizingSum ]
  ifFalse: [ 1.0 / Action size ]
  ]
```

Ah, this is just bad. It is awful.

Using normalized on Regret unwraps it.

I just get the sense I do not understand the idiomatic OO. But pattern matching or not, I can make my usual style work and write Pharo code in it. I gave what I thought should be done in Pharo a try and it is just not going well.

The separation between data and and functions operating on data is a fundamental one in my mind. I do not think that sending messages to instances is the best way to go, if I want to make this work, I'll lean more on metaclass methods.

When I do it to completeness I'll ask around for a code review.

7:20pm. Damn I am tired. I am done for the day here."