"6:45am. I am up. Today is an example of a day when I get up really early and do not manage to fall back asleep. Yawn. Damn it is freezing here in this room. It sucks not being to fall asleep at will.
https://mega.nz/#F!jgFkzaSA!5GI_Z436e98vuY2BVbhtJQ!bwsG1SLS

I got these English Asuka rips from yesterday's thread so I might go for that first.

That having said, maybe the lack of sleep is caused by me not being able to work through some design issues in Pharo.

Let me take the `normalize` function as an example.

Everything in Pharo works on instances. In theory there should not be that much difference between instance methods and static methods apart from an extra argument in the later, but Pharo has no thing such as the pipe operator and its whole grammar is structured towards a particular kind of style.

So I need to call `normalize` on an instance, in this case of an array. So far so good.

But where exactly should I put this? Directly into the array class?

This feels wrong - when it comes time to push this to Github how exactly will things go? Putting it directly into the array class feels like it will turn into a versioning nightmare eventually.

So I guess the only choice is to inherit from it. This might be fine for the problem I am doing now, but later I want to do a different problem and do not want to share implementations.

Before I continue, I need to ask about this. Let me first make the normalize method and test it out before I post it on Github.

I really, really like how Pharo looks, but I am 2 days into it and already running into problems. I guess it just goes side by side with functions being the ideal unit of abstraction. My ideal style is really the inverse of what Pharo offers. But I want to give it a honest effort for a while longer. I can't give up just here.

7:15am. Well, let me do it. I'll chill a bit after that. I'll post the question on SO. I do not feel like messing with Discord."